@@ -2,8 +2,8 @@
 
 namespace App\Analysis\LimitAnalysis;
 
+use App\Air\Measurement\MeasurementInterface;
 use App\Entity\Station;
-use App\Pollution\Pollutant\PollutantInterface;
 use FOS\ElasticaBundle\Finder\PaginatedFinderInterface;
 
 class LimitAnalysis implements LimitAnalysisInterface
@@ -11,8 +11,8 @@ class LimitAnalysis implements LimitAnalysisInterface
     /** @var Station $station */
     protected $station;
 
-    /** @var PollutantInterface $pollutant */
-    protected $pollutant;
+    /** @var MeasurementInterface $measurement */
+    protected $measurement;
 
     /** @var PaginatedFinderInterface $finder */
     protected $finder;
@@ -35,9 +35,9 @@ class LimitAnalysis implements LimitAnalysisInterface
         return $this;
     }
 
-    public function setPollutant(PollutantInterface $pollutant): LimitAnalysisInterface
+    public function setMeasurement(MeasurementInterface $measurement): LimitAnalysisInterface
     {
-        $this->pollutant = $pollutant;
+        $this->measurement = $measurement;
 
         return $this;
     }
