@@ -68,7 +68,7 @@ class LuftdatenArchiveCommand extends ContainerAwareCommand
             });
 
             foreach ($valueList as $value) {
-                $this->getContainer()->get('old_sound_rabbit_mq.value_producer')->publish(serialize($value));
+                $this->getContainer()->get('old_sound_rabbit_mq.luft_value_producer')->publish(serialize($value));
             }
 
             $counter += count($valueList);
