@@ -6,7 +6,9 @@ use App\Entity\TwitterSchedule;
 use App\Twitter\TwitterInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Helper\Table;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 class TweetCommand extends Command
@@ -25,13 +27,18 @@ class TweetCommand extends Command
         $this
             ->setName('luft:tweet')
             ->setDescription('Post current data')
-            ->addOption('dry-run');
+            ->addOption('dry-run')
+            ->addOption('dateTime', null, InputOption::VALUE_REQUIRED);
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         if ($input->getOption('dry-run')) {
             $this->twitter->setDryRun(true);
+        }
+
+        if ($input->getOption('dateTime')) {
+            $this->twitter->setDateTime(new \Datetime($input->getOption('dateTime')));
         }
 
         $this->twitter->tweet();
