@@ -4,11 +4,19 @@ namespace App\Controller;
 
 use App\Analysis\FireworksAnalysis\FireworksAnalysisInterface;
 use App\Analysis\KomfortofenAnalysis\KomfortofenAnalysisInterface;
+use App\SeoPage\SeoPageInterface;
 use Symfony\Component\HttpFoundation\Response;
+use Flagception\Bundle\FlagceptionBundle\Annotations\Feature;
 
+/**
+ * @Feature("analysis")
+ */
 class AnalysisController extends AbstractController
 {
-    public function komfortofenAction(KomfortofenAnalysisInterface $komfortofenAnalysis): Response
+    /**
+     * @Feature("analysis_komfortofen")
+     */
+    public function komfortofenAction(KomfortofenAnalysisInterface $komfortofenAnalysis, SeoPageInterface $seoPage): Response
     {
         $interval = new \DateInterval('P30D');
         $untilDateTime = new \DateTimeImmutable();
@@ -20,13 +28,24 @@ class AnalysisController extends AbstractController
 
         $ofens = $komfortofenAnalysis->analyze();
 
+        $seoPage->setTitle('Finde Feinstaub-Belastungen aus Holzöfen in deiner Nähe');
+
         return $this->render('Analysis/komfortofen.html.twig', [
             'ofenList' => $ofens,
         ]);
     }
 
-    public function fireworksAction(FireworksAnalysisInterface $fireworksAnalysis): Response
+    /**
+     * @Feature("analysis_fireworks")
+     */
+    public function fireworksAction(FireworksAnalysisInterface $fireworksAnalysis, SeoPageInterface $seoPage): Response
     {
+        $seoPage
+            ->setTwitterPreviewPhoto('/img/share/silvester/twitter.jpg')
+            ->setOpenGraphPreviewPhoto('/img/share/silvester/facebook.jpg')
+            ->setTitle('Feinstaub aus Silvester-Feuerwerken')
+            ->setDescription('Finde erhöhte Feinstaub-Konzentrationen aus Silvester-Feuerwerken');
+
         $fireworksAnalysis = $fireworksAnalysis->analyze();
 
         return $this->render('Analysis/fireworks.html.twig', [
