@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\City;
 use App\Entity\Data;
 use App\Entity\Station;
+use App\Geocoding\RequestConverter\RequestConverterInterface;
 use App\Pollution\PollutionDataFactory\PollutionDataFactory;
 use JMS\Serializer\SerializerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -26,7 +27,7 @@ class ApiController extends AbstractController
      *   description="Retrieve pollution data for station",
      *   @SWG\Schema(
      *     type="array",
-     *     @SWG\Items(ref=@Model(type=App\Pollution\Box\Box::class))
+     *     @SWG\Items(ref=@Model(type=App\Air\ViewModel\MeasurementViewModel::class))
      *   )
      * )
      * @SWG\Parameter(
@@ -47,7 +48,7 @@ class ApiController extends AbstractController
             throw $this->createNotFoundException();
         }
 
-        $pollutantList = $pollutionDataFactory->setCoord($station)->createDecoratedPollutantList();
+        $pollutantList = $pollutionDataFactory->setStation($station)->createDecoratedPollutantList();
 
         return new JsonResponse($serializer->serialize($this->unpackPollutantList($pollutantList), 'json'), 200, [], true);
     }
@@ -61,7 +62,7 @@ class ApiController extends AbstractController
      *   description="Retrieve pollution data for cities",
      *   @SWG\Schema(
      *     type="array",
-     *     @SWG\Items(ref=@Model(type=App\Pollution\Box\Box::class))
+     *     @SWG\Items(ref=@Model(type=App\Air\ViewModel\MeasurementViewModel::class))
      *   )
      * )
      */
@@ -77,9 +78,9 @@ class ApiController extends AbstractController
         }
 
         $stationList = $this->getStationListForCity($city);
-        $stationsBoxList = $this->createBoxListForStationList($pollutionDataFactory, $stationList);
+        $stationViewModelList = $this->createViewModelListForStationList($pollutionDataFactory, $stationList);
 
-        return new JsonResponse($serializer->serialize($stationsBoxList, 'json'), 200, [], true);
+        return new JsonResponse($serializer->serialize($stationViewModelList, 'json'), 200, [], true);
     }
 
     /**
@@ -109,16 +110,17 @@ class ApiController extends AbstractController
      *   description="Returns pollution data of specified station",
      *   @SWG\Schema(
      *     type="array",
-     *     @SWG\Items(ref=@Model(type=App\Pollution\Box\Box::class))
+     *     @SWG\Items(ref=@Model(type=App\Air\ViewModel\MeasurementViewModel::class))
      *   )
      * )
      */
     public function displayAction(
         Request $request,
         SerializerInterface $serializer,
-        PollutionDataFactory $pollutionDataFactory
+        PollutionDataFactory $pollutionDataFactory,
+        RequestConverterInterface $requestConverter
     ): Response {
-        $coord = $this->getCoordByRequest($request);
+        $coord = $requestConverter->getCoordByRequest($request);
 
         if (!$coord) {
             throw $this->createNotFoundException();
@@ -182,8 +184,10 @@ class ApiController extends AbstractController
      *   @Model(type=App\Entity\Station::class)
      * )
      */
-    public function stationAction(SerializerInterface $serializer, string $stationCode = null): Response
+    public function stationAction(Request $request, SerializerInterface $serializer, string $stationCode = null): Response
     {
+        $providerIdentifier = $request->get('provider');
+
         if ($stationCode) {
             $station = $this->getDoctrine()->getRepository(Station::class)->findOneByStationCode($stationCode);
 
@@ -192,6 +196,8 @@ class ApiController extends AbstractController
             }
 
             return new JsonResponse($serializer->serialize($station, 'json'), 200, [], true);
+        } elseif ($providerIdentifier) {
+            $stationList = $this->getDoctrine()->getRepository(Station::class)->findActiveStationsByProvider($providerIdentifier);
         } else {
             $stationList = $this->getDoctrine()->getRepository(Station::class)->findAll();
         }
@@ -217,12 +223,12 @@ class ApiController extends AbstractController
 
     protected function unpackPollutantList(array $pollutantList): array
     {
-        $boxList = [];
+        $viewModelList = [];
 
         foreach ($pollutantList as $pollutant) {
-            $boxList = array_merge($boxList, $pollutant);
+            $viewModelList = array_merge($viewModelList, $pollutant);
         }
 
-        return $boxList;
+        return $viewModelList;
     }
 }
