@@ -4,16 +4,21 @@ namespace App\Controller;
 
 use App\Analysis\LimitAnalysis\LimitAnalysisInterface;
 use App\Entity\Station;
+use App\Plotter\StationPlotter\StationPlotterInterface;
 use App\Pollution\PollutionDataFactory\HistoryDataFactoryInterface;
 use App\Pollution\PollutionDataFactory\PollutionDataFactory;
-use App\SeoPage\SeoPage;
+use App\SeoPage\SeoPageInterface;
 use App\Util\DateTimeUtil;
+use Flagception\Bundle\FlagceptionBundle\Annotations\Feature;
+use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\RouterInterface;
+use WhiteOctober\BreadcrumbsBundle\Model\Breadcrumbs;
 
 class StationController extends AbstractController
 {
-    public function stationAction(SeoPage $seoPage, string $stationCode, PollutionDataFactory $pollutionDataFactory): Response
+    public function stationAction(SeoPageInterface $seoPage, string $stationCode, PollutionDataFactory $pollutionDataFactory, Breadcrumbs $breadcrumbs, RouterInterface $router): Response
     {
         /** @var Station $station */
         $station = $this->getDoctrine()->getRepository(Station::class)->findOneByStationCode($stationCode);
@@ -22,19 +27,24 @@ class StationController extends AbstractController
             throw $this->createNotFoundException();
         }
 
-        $boxList = $pollutionDataFactory
+        $viewModelList = $pollutionDataFactory
             ->setStation($station)
             ->createDecoratedPollutantList();
 
         if ($station->getCity()) {
             $seoPage->setTitle(sprintf('Luftmesswerte für die Station %s — Feinstaub, Stickstoffdioxid und Ozon in %s', $station->getStationCode(), $station->getCity()->getName()));
+
+            $breadcrumbs
+                ->addItem('Luft', $router->generate('display'))
+                ->addItem($station->getCity()->getName(), $router->generate('show_city', ['citySlug' => $station->getCity()->getSlug()]))
+                ->addItem($station->getStationCode(), $router->generate('station', ['stationCode' => $station->getStationCode()]));
         } else {
             $seoPage->setTitle(sprintf('Luftmesswerte für die Station %s', $station->getStationCode()));
         }
 
         return $this->render('Default/station.html.twig', [
             'station' => $station,
-            'pollutantList' => $boxList,
+            'pollutantList' => $viewModelList,
         ]);
     }
 
@@ -62,7 +72,10 @@ class StationController extends AbstractController
         ]);
     }
 
-    public function historyAction(Request $request, string $stationCode, HistoryDataFactoryInterface $historyDataFactory): Response
+    /**
+     * @Feature("station_history")
+     */
+    public function historyAction(Request $request, string $stationCode, SeoPageInterface $seoPage, HistoryDataFactoryInterface $historyDataFactory, RouterInterface $router): Response
     {
         if ($untilDateTimeParam = $request->query->get('until')) {
             try {
@@ -93,6 +106,20 @@ class StationController extends AbstractController
             throw $this->createNotFoundException();
         }
 
+        if ($station->getCity()) {
+            $seoPage->setTitle(sprintf('Frühere Luftmesswerte für die Station %s — Feinstaub, Stickstoffdioxid und Ozon in %s', $station->getStationCode(), $station->getCity()->getName()));
+        } else {
+            $seoPage->setTitle(sprintf('Frühere Luftmesswerte für die Station %s', $station->getStationCode()));
+        }
+
+        $seoPage->setOpenGraphPreviewPhoto($router->generate('station_history_plot', [
+            'stationCode' => $station->getStationCode(),
+            'from' => $fromDateTime->format('Y-m-d'),
+            'until' => $untilDateTime->format('Y-m-d'),
+            'width' => 1200,
+            'height' => 630,
+        ]));
+
         $dataLists = $historyDataFactory
             ->setStation($station)
             ->createDecoratedPollutantListForInterval($fromDateTime, $untilDateTime);
@@ -117,5 +144,60 @@ class StationController extends AbstractController
         }
 
         return array_unique($pollutantIdList);
+    }
+
+    /**
+     * @Feature("station_history")
+     */
+    public function plotHistoryAction(Request $request, string $stationCode, StationPlotterInterface $stationPlotter, string $graphCacheDirectory): BinaryFileResponse
+    {
+        if ($untilDateTimeParam = $request->query->get('until')) {
+            try {
+                $untilDateTime = DateTimeUtil::getDayEndDateTime(new \DateTime($untilDateTimeParam));
+            } catch (\Exception $exception) {
+                $untilDateTime = DateTimeUtil::getHourStartDateTime(new \DateTime());
+            }
+        } else {
+            $untilDateTime = DateTimeUtil::getHourStartDateTime(new \DateTime());
+        }
+
+        if ($fromDateTimeParam = $request->query->get('from')) {
+            try {
+                $fromDateTime = DateTimeUtil::getDayStartDateTime(new \DateTime($fromDateTimeParam));
+            } catch (\Exception $exception) {
+                $fromDateTime = DateTimeUtil::getHourStartDateTime(new \DateTime());
+                $fromDateTime->sub(new \DateInterval('P3D'));
+            }
+        } else {
+            $fromDateTime = DateTimeUtil::getHourStartDateTime(new \DateTime());
+            $fromDateTime->sub(new \DateInterval('P3D'));
+        }
+
+        /** @var Station $station */
+        $station = $this->getDoctrine()->getRepository(Station::class)->findOneByStationCode($stationCode);
+
+        if (!$station) {
+            throw $this->createNotFoundException();
+        }
+
+        $width = (int) $request->get('width', 800);
+        $height = (int) $request->get('height', 400);
+
+        $filename = sprintf('%s/%s-%d-%d-%dx%d.png', $graphCacheDirectory, $station->getStationCode(), $fromDateTime->format('U'), $untilDateTime->format('U'), $width, $height);
+
+        if (!file_exists($filename)) {
+            $stationPlotter
+                ->setWidth($width)
+                ->setHeight($height)
+                ->setTitle(sprintf('Messwerte der Station %s', $station->getStationCode()))
+                ->setStation($station)
+                ->setFromDateTime($fromDateTime)
+                ->setUntilDateTime($untilDateTime)
+                ->plot($filename);
+        }
+
+        $response = new BinaryFileResponse($filename);
+
+        return $response;
     }
 }
