@@ -2,6 +2,7 @@
 
 namespace App\Pollution\PollutantFactoryStrategy;
 
+use App\Air\Measurement\MeasurementInterface;
 use App\Entity\Data;
 use App\Pollution\DataList\DataListInterface;
 
@@ -13,7 +14,15 @@ class LuftdatenPollutantFactoryStrategy implements PollutantFactoryStrategyInter
         $missingList = [];
 
         array_walk($list, function(array $list, int $key) use (&$missingList) {
-            if (null === $list || 0 === count($list) || 1 === count($list)) {
+            if (null === $list) {
+                array_push($missingList, $key);
+            }
+
+            if (in_array($key, [MeasurementInterface::MEASUREMENT_PM25, MeasurementInterface::MEASUREMENT_PM10]) && count($list) < 2) {
+                array_push($missingList, $key);
+            }
+
+            if (!in_array($key, [MeasurementInterface::MEASUREMENT_PM25, MeasurementInterface::MEASUREMENT_PM10]) && count($list) < 1) {
                 array_push($missingList, $key);
             }
         });
@@ -38,6 +47,25 @@ class LuftdatenPollutantFactoryStrategy implements PollutantFactoryStrategyInter
         $existingData = $this->getExistingSingleData($dataList, $data->getPollutant());
 
         return $this->isProvidersDifferent($existingData, $data);
+    }
+
+    public function isSatisfied(DataListInterface $dataList, int $pollutantId): bool
+    {
+        // Luftdaten only collects pm10 and pm25, so there is no need to wait for o3, no2 or so2
+        if (in_array($pollutantId, [MeasurementInterface::MEASUREMENT_PM10, MeasurementInterface::MEASUREMENT_PM25])) {
+            if ($dataList->countPollutant($pollutantId) === 2) {
+                $list = $dataList->getList()[$pollutantId];
+
+                $data1 = array_pop($list);
+                $data2 = array_pop($list);
+
+                return $this->isProvidersDifferent($data1, $data2);
+            }
+
+            return false;
+        }
+
+        return $dataList->countPollutant($pollutantId) >= 1;
     }
 
     public function addDataToList(DataListInterface $dataList, Data $data = null): bool
