@@ -2,9 +2,8 @@
 
 namespace App\Pollution\PollutionDataFactory;
 
+use App\Air\ViewModelFactory\MeasurementViewModelFactoryInterface;
 use App\Entity\Station;
-use App\Pollution\Box\Box;
-use App\Pollution\BoxDecorator\BoxDecoratorInterface;
 use App\Pollution\DataList\DataList;
 use App\Pollution\DataRetriever\DataRetrieverInterface;
 use App\Pollution\PollutantFactoryStrategy\PollutantFactoryStrategyInterface;
@@ -19,26 +18,23 @@ abstract class AbstractPollutionDataFactory implements PollutionDataFactoryInter
     /** @var StationFinderInterface $stationFinder */
     protected $stationFinder;
 
-    /** @var StationFinderInterface $boxDecorator */
-    protected $boxDecorator;
+    /** @var MeasurementViewModelFactoryInterface $viewModelFactory */
+    protected $measurementViewModelFactory;
 
     /** @var DataList $dataList */
     protected $dataList;
 
-    /** @var DataRetrieverInterface */
+    /** @var DataRetrieverInterface $dataRetriever*/
     protected $dataRetriever;
-
-    /** @var array $stationList */
-    protected $stationList = [];
 
     /** @var PollutantFactoryStrategyInterface $strategy */
     protected $strategy;
 
-    public function __construct(StationFinderInterface $stationFinder, BoxDecoratorInterface $boxDecorator, DataRetrieverInterface $dataRetriever, PollutantFactoryStrategyInterface $strategy)
+    public function __construct(StationFinderInterface $stationFinder, MeasurementViewModelFactoryInterface $viewModelFactory, DataRetrieverInterface $dataRetriever, PollutantFactoryStrategyInterface $strategy)
     {
         $this->stationFinder = $stationFinder;
         $this->dataList = new DataList();
-        $this->boxDecorator = $boxDecorator;
+        $this->measurementViewModelFactory = $viewModelFactory;
         $this->dataRetriever = $dataRetriever;
         $this->strategy = $strategy;
     }
@@ -46,7 +42,6 @@ abstract class AbstractPollutionDataFactory implements PollutionDataFactoryInter
     public function setCoord(CoordInterface $coord): PollutionDataFactoryInterface
     {
         $this->coord = $coord;
-        $this->stationList = $this->stationFinder->setCoord($this->coord)->findNearestStations();
 
         return $this;
     }
@@ -54,7 +49,6 @@ abstract class AbstractPollutionDataFactory implements PollutionDataFactoryInter
     public function setStation(Station $station): PollutionDataFactoryInterface
     {
         $this->coord = $station;
-        $this->stationList = [$station];
 
         return $this;
     }
