@@ -2,36 +2,65 @@
 
 namespace App\Pollution\PollutionDataFactory;
 
+use App\Entity\Data;
+use App\Util\DateTimeUtil;
+
 class HistoryDataFactory extends PollutionDataFactory implements HistoryDataFactoryInterface
 {
     public function createDecoratedPollutantListForInterval(\DateTime $fromDateTime, \DateTime $untilDateTime): array
     {
         $dataLists = $this->getDataListsForInterval($fromDateTime, $untilDateTime);
 
-        $boxLists = [];
+        $dataLists = $this->convert($dataLists);
+
+        $measurementModelLists = [];
 
         /** @var array $dataList */
         foreach ($dataLists as $timestamp => $dataList) {
-            $boxList = $this->getBoxListFromDataList($dataList);
-            $boxLists[$timestamp] = $this->decoratePollutantList($boxList);
+            $measurementModelList = $this->getMeasurementViewModelListFromDataList($dataList);
+
+            $measurementModelLists[$timestamp] = $this->decoratePollutantList($measurementModelList);
         }
 
-        return $boxLists;
+        return $measurementModelLists;
     }
 
     public function getDataListsForInterval(\DateTime $fromDateTime, \DateTime $untilDateTime): array
     {
         $dataListList = [];
 
-        $hour = new \DateInterval('PT1H');
-        $dateTime = clone $fromDateTime;
+        $diffInterval = $fromDateTime->diff($untilDateTime);
 
-        while ($dateTime < $untilDateTime) {
-            $dataListList[$dateTime->format('U')] = $this->getDataListFromStationList($this->stationList, $dateTime, $hour, 'ASC');
+        $this->dataList->reset();
 
-            $dateTime->add($hour);
+        $missingPollutants = $this->strategy->getMissingPollutants($this->dataList);
+
+        foreach ($missingPollutants as $pollutantId) {
+            $dataListList[$pollutantId] = $this->dataRetriever->retrieveDataForCoord($this->coord, $pollutantId, $fromDateTime, $diffInterval);
         }
 
         return $dataListList;
+    }
+
+    protected function convert(array $dataListLists): array
+    {
+        $newDataListLists = [];
+
+        foreach ($dataListLists as $pollutantId => $dataList) {
+            /** @var Data $data */
+            foreach ($dataList as $data) {
+                $timestamp = DateTimeUtil::getHourStartDateTime($data->getDateTime())->format('U');
+
+                if (!array_key_exists($timestamp, $newDataListLists)) {
+                    $newDataListLists[$timestamp] = [$pollutantId => [$data]];
+                } elseif (!array_key_exists($pollutantId, $newDataListLists[$timestamp])) {
+                    $newDataListLists[$timestamp][$pollutantId] = [$data];
+                } elseif ($newDataListLists[$timestamp][$pollutantId][0]->getValue() < $data->getValue()) {
+                    $newDataListLists[$timestamp][$pollutantId][0] = $data;
+                }
+            }
+        }
+
+        return $newDataListLists;
     }
 }
