@@ -2,71 +2,80 @@
 
 namespace App\Pollution\PollutionDataFactory;
 
+use App\Air\ViewModel\MeasurementViewModel;
 use App\Entity\Data;
-use App\Entity\Station;
-use App\Pollution\Box\Box;
+use App\Pollution\UniqueStrategy\Hasher;
 
 class PollutionDataFactory extends AbstractPollutionDataFactory
 {
-    public function createDecoratedPollutantList(): array
+    public function createDecoratedPollutantList(\DateTime $dateTime = null, \DateInterval $dateInterval = null, int $workingSetSize = 20): array
     {
-        $dataList = $this->getDataListFromStationList($this->stationList);
+        if (!$dateTime) {
+            $dateTime = new \DateTime();
+        }
 
-        $boxList = $this->getBoxListFromDataList($dataList);
+        if (!$dateInterval) {
+            $dateInterval = new \DateInterval('PT12H');
+        }
 
-        $boxList = $this->decoratePollutantList($boxList);
+        $dateTime->sub($dateInterval);
 
-        return $boxList;
+        $dataList = $this->getDataListFromStationList($dateTime, $dateInterval, $workingSetSize);
+
+        $measurementViewModelList = $this->getMeasurementViewModelListFromDataList($dataList);
+
+        $measurementViewModelList = $this->decoratePollutantList($measurementViewModelList);
+
+        return $measurementViewModelList;
     }
 
-    protected function getDataListFromStationList(array $stationList, \DateTime $dateTime = null, \DateInterval $interval = null): array
+    protected function getDataListFromStationList(\DateTime $fromDateTime = null, \DateInterval $interval = null, int $workingSetSize): array
     {
         $this->dataList->reset();
 
         $missingPollutants = $this->strategy->getMissingPollutants($this->dataList);
 
-        /** @var Station $station */
-        foreach ($stationList as $station) {
-            foreach ($missingPollutants as $pollutant) {
-                $data = $this->dataRetriever->retrieveStationData($station, $pollutant, $dateTime, $interval);
+        foreach ($missingPollutants as $pollutantId) {
+            $dataList = $this->dataRetriever->retrieveDataForCoord($this->coord, $pollutantId, $fromDateTime, $interval, 20.0, $workingSetSize);
+
+            if (0 === count($dataList)) {
+                continue;
+            }
+
+            while (!$this->strategy->isSatisfied($this->dataList, $pollutantId) && count($dataList)) {
+                $data = array_shift($dataList);
 
                 if ($this->strategy->accepts($this->dataList, $data)) {
                     $this->strategy->addDataToList($this->dataList, $data);
                 }
             }
-
-            $missingPollutants = $this->strategy->getMissingPollutants($this->dataList);
-
-            if (0 === count($missingPollutants)) {
-                break;
-            }
         }
-
+        
         return $this->dataList->getList();
     }
 
-    protected function getBoxListFromDataList(array $dataList): array
+    protected function getMeasurementViewModelListFromDataList(array $dataList): array
     {
-        $pollutantList = [];
+        $measurementViewModelList = [];
 
         /** @var array $data */
         foreach ($dataList as $data) {
             /** @var Data $dataElement */
             foreach ($data as $dataElement) {
                 if ($dataElement) {
-                    $pollutantList[$dataElement->getPollutant()][$dataElement->getId()] = new Box($dataElement);
+                    $measurementViewModelList[$dataElement->getPollutant()][Hasher::hashData($dataElement)] = new MeasurementViewModel($dataElement);
                 }
             }
         }
 
-        return $pollutantList;
+        return $measurementViewModelList;
     }
 
     protected function decoratePollutantList(array $pollutantList): array
     {
         return $this
             ->reset()
-            ->boxDecorator
+            ->measurementViewModelFactory
             ->setCoord($this->coord)
             ->setPollutantList($pollutantList)
             ->decorate()
