@@ -7,10 +7,9 @@ use App\Provider\UmweltbundesamtDe\SourceFetcher\Reporting\ReportingInterface;
 
 class UbaCOQuery extends AbstractUbaQuery
 {
-    public function __construct(ReportingInterface $reporting)
-    {
-        parent::__construct($reporting);
+    /** @var int $component */
+    protected $component = 2;
 
-        $this->filter = new COFilter();
-    }
+    /** @var array $scope */
+    protected $scope = [4, 5];
 }
