@@ -2,14 +2,17 @@
 
 namespace App\Provider\UmweltbundesamtDe\SourceFetcher\Query;
 
-use App\Provider\UmweltbundesamtDe\SourceFetcher\Filter\FilterInterface;
-use App\Provider\UmweltbundesamtDe\SourceFetcher\Reporting\ReportingInterface;
-
 interface UbaQueryInterface
 {
-    public function getQueryString(): string;
-    public function getQueryOptions(): array;
-    public function getDateTimeFormat(): string;
-    public function getReporting(): ReportingInterface;
-    public function getFilter(): FilterInterface;
+    public function getComponent(): int;
+
+    public function getScope(): array;
+
+    public function getFromDateTime(): \DateTime;
+
+    public function setFromDateTime(\DateTime $fromDateTime): UbaQueryInterface;
+
+    public function getUntilDateTime(): \DateTime;
+
+    public function setUntilDateTime(\DateTime $untilDateTime): UbaQueryInterface;
 }
