@@ -2,8 +2,7 @@
 
 namespace App\Twitter\MessageFactory;
 
-use App\Pollution\Box\Box;
-use App\Pollution\PollutionLevel\PollutionLevel;
+use App\Air\ViewModel\MeasurementViewModel;
 use Symfony\Component\Translation\TranslatorInterface;
 
 class EmojiMessageFactory extends AbstractMessageFactory
@@ -22,9 +21,9 @@ class EmojiMessageFactory extends AbstractMessageFactory
 
         /** @var array $pollutant */
         foreach ($this->pollutantList as $pollutant) {
-            /** @var Box $box */
-            foreach ($pollutant as $box) {
-                $this->message .= sprintf("%s %s: %.0f %s \n", $this->getEmoji($box), $box->getPollutant()->getName(), $box->getData()->getValue(), $box->getPollutant()->getUnitPlain());
+            /** @var MeasurementViewModel $measurementViewModel */
+            foreach ($pollutant as $measurementViewModel) {
+                $this->message .= sprintf("%s %s: %.0f %s \n", $this->getEmoji($measurementViewModel), $measurementViewModel->getPollutant()->getName(), $measurementViewModel->getData()->getValue(), $measurementViewModel->getPollutant()->getUnitPlain());
             }
         }
 
@@ -33,9 +32,9 @@ class EmojiMessageFactory extends AbstractMessageFactory
         return $this;
     }
 
-    protected function getEmoji(Box $box): string
+    protected function getEmoji(MeasurementViewModel $measurementViewModel): string
     {
-        $translationKey = sprintf('air_quality.index.%d.icon', $box->getPollutionLevel());
+        $translationKey = sprintf('air_quality.index.%d.icon', $measurementViewModel->getPollutionLevel());
 
         return $this->translator->trans($translationKey);
     }
